Fixes webpack-dev-server issues.
There were issues with the webpack dev server that had to do with how
the webpack config file was setup. This commit fixes the webpack config
so the webpack dev server will now actually hot-reload upon source
changes.

This commit also reorganizes the sass files and the output locations of
both the index bundle and the compiled css file to the root directory.
In the future it may be worth refactoring the project so that all
bundled assets appear once again in dist/, but for now this is working
fine.